About this commit:
1. server/builds.go:92 uses SetStatus().
2. The other APIs we talked about: Status(), StatusList(), we were able to move out.
3. The repo_del_test.go code merge into repo_test.go
4. Unit tests for the other build APIs added.
   - We are facing a crash in: github.com/drone/drone/datastore/bolt.(*DB).SetBuildTask(0xc208056080, 0x5e15d0, 0x15, 0x1, 0xc208036940, 0x0, 0x0)
     which seems to be related to the note in: build.go:207 (// TODO check index to prevent nil pointer / panic)
5. With these new tests we get over 86% plus test cover for bolt package.
