+++ datastore/datastore.go
@@ -108,11 +108,11 @@ type Datastore interface {
 
 	// Status returns the status for the given repository
 	// and build number.
-	Status(string, int, string) (*common.Status, error)
+	////Status(string, int, string) (*common.Status, error)
 
 	// StatusList returned a list of all build statues for
 	// the given repository and build number.
-	StatusList(string, int) ([]*common.Status, error)
+	////StatusList(string, int) ([]*common.Status, error)
 
 	// SetStatus inserts a new build status for the
 	// named repository and build number. If the status already
